backend/requirements.txt
Django>=4.2,<5.0
requests
pandas
yfinance
python-decouple
